+++ payments/src/funcs/bulkOperationsDeleteCollaborator.ts
@@ -90,6 +90,7 @@ export async function bulkOperationsDeleteCollaborator(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "DELETE",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,