+++ payments/src/funcs/bulkOperationsGetSpecific.ts
@@ -94,6 +94,7 @@ export async function bulkOperationsGetSpecific(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,