+++ payments/src/funcs/bulkOperationsUpdateClient.ts
@@ -88,6 +88,7 @@ export async function bulkOperationsUpdateClient(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "PATCH",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,