+++ payments/src/funcs/clientManagementDeleteClientDeduction.ts
@@ -88,6 +88,7 @@ export async function clientManagementDeleteClientDeduction(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "DELETE",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,