+++ payments/src/funcs/collaboratorManagementGetEventAssociated.ts
@@ -90,6 +90,7 @@ export async function collaboratorManagementGetEventAssociated(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,