+++ payments/src/funcs/deprecatedUpdateInstantPayout.ts
@@ -85,6 +85,7 @@ export async function deprecatedUpdateInstantPayout(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,