+++ payments/src/funcs/documentSigningAndEligibilityUpdate.ts
@@ -95,6 +95,7 @@ export async function documentSigningAndEligibilityUpdate(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "PATCH",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,