+++ payments/src/funcs/oneThousandAndNinetyNineOperationsCreateItem.ts
@@ -95,6 +95,7 @@ export async function oneThousandAndNinetyNineOperationsCreateItem(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,