+++ payments/src/funcs/oneThousandAndNinetyNineOperationsGetForm1099.ts
@@ -97,6 +97,7 @@ export async function oneThousandAndNinetyNineOperationsGetForm1099(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,