+++ payments/src/funcs/oneThousandAndNinetyNineOperationsListItem.ts
@@ -92,6 +92,7 @@ export async function oneThousandAndNinetyNineOperationsListItem(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,