+++ payments/src/funcs/wingspanWalletCreate.ts
@@ -84,6 +84,7 @@ export async function wingspanWalletCreate(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,